Add files via upload
This code was difficult for me as a beginner, but it makes sense after hours of staring at it. I had to look elsewhere for help such as the web and YouTube. I used the function getAnIntFromTheUser and compareTwoInts to compare both integers and lastly sumTwoInts to sum both of the integers. I tested it out with different numbers and it ran the code the way I wanted it to.